Rename App to SwapAppV1
