+++ src/apps/swap-v1/index.tsx
@@ -3,7 +3,7 @@ import 'react-hot-loader'
 import React from 'react'
 import ReactDOM from 'react-dom'
 
-import App from 'apps/swap-v1/App'
+import SwapAppV1 from 'apps/swap-v1/SwapAppV1'
 
 const root = document.getElementById('root')
-ReactDOM.render(<App />, root)
+ReactDOM.render(<SwapAppV1 />, root)
